Fixed bug

+++ Assets/Scripts/Server/Master/RequestReceiver.cs
@@ -100,10 +100,10 @@ public class RequestReceiver : MonoBehaviour
             int userId = (int)data[1];
             if (RoomManager.Instance.JoinMinigameRoom(roomCode, userId))
             {
-                SendReply.Instance.SendJoinRoomReply(userId, RoomManager.Instance.roomInfos[roomCode].minigame);
+                SendReply.Instance.SendJoinRoomReply(userId, (int)RoomManager.Instance.roomInfos[roomCode].minigame);
             } else
             {
-                SendReply.Instance.SendJoinRoomReply(userId, Minigame.None);
+                SendReply.Instance.SendJoinRoomReply(userId, (int)Minigame.None);
             }
         }
     }
